requirements.txt
amqp==2.6.1
asgiref==3.6.0
billiard==3.6.4.0
ccard==0.1.9
celery==4.4.7
certifi==2022.12.7
charset-normalizer==2.1.1
cloudinary==1.30.0
Django==3.2.7
django-celery-beat==2.1.0
django-cloudinary-storage==0.3.0
django-timezone-field==4.2.3
idna==3.4
kombu==4.6.11
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7
redis==3.5.3
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
vine==1.3.0
